removed code of public token
public token is for viewing model on browser. It derives from the code of HubBrowser tutorial. In this version of this Issue API sample, we have not section on viewing model. So removed the relevant code to avoid any confusing.

It might be added back when ACC pushpin (markup) API is supported, where we will demo the workflow of loading issue in viewer.